r100: fixed a nasty bug caused by kstring_t
In the new kseq.h, we have "struct{size_t l,m}kstring_t", but in the kstring.h,
we have "struct{uint32_t l,m}kstring_t". In seq.c where kseq is instantiated,
kstring_t in kstring.h overrides kseq.h, which makes the "kstring_t" type
different and causes segfault.

In the old kseq.h, we have "struct{unsigned l,m}kstring_t", which is fine.

+++ kseq.h
@@ -81,7 +81,7 @@
 #ifndef KSTRING_T
 #define KSTRING_T kstring_t
 typedef struct __kstring_t {
-	size_t l, m;
+	unsigned l, m;
 	char *s;
 } kstring_t;
 #endif
